[Feat] 클래스 추가 및 필드 변경, 복합키 구성 (#1)
- 버전, 약관id를 가진 subterm 클래스 추가
- 기존 term 클래스.detail 필드 제거
- 기존 term 클래스.detail 필드를 사용하는 service, repository 내용 삭제 및 null로 변경
- 유저가 동의한 약관 목록을 저장하는 userSubTerm 클래스 추가
- userSubTerm 클래스.버전/약관id/유저id를 갖도록 수정
- @EmbeddedId, @Embeddable를 사용해 복합키 구성, @MapsId를 통한 외래키 매핑
- @GeneratedValue 매핑전략으로 Identity 사용
