--- src/main/java/com/flab/s_market/delivery/domain/Delivery.java
@@ -1,25 +1,0 @@
-package com.flab.s_market.delivery.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Delivery {
-    @Id @GeneratedValue
-    @Column(name = "delivery_id")
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private DeliveryStatus status;
-}
