--- src/main/java/com/flab/s_market/delivery/domain/DeliveryStatus.java
@@ -1,5 +1,0 @@
-package com.flab.s_market.delivery.domain;
-
-public enum DeliveryStatus {
-    PREPARE_PRODUCT, DELIVER, COMPLETE
-}
