+++ src/main/java/com/flab/s_market/domains/delivery/domain/DeliveryStatus.java
@@ -1,0 +1,5 @@
+package com.flab.s_market.domains.delivery.domain;
+
+public enum DeliveryStatus {
+    PREPARE_PRODUCT, DELIVER, COMPLETE
+}
