+++ src/main/java/com/flab/s_market/domains/event/domain/Event.java
@@ -1,0 +1,31 @@
+package com.flab.s_market.domains.event.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Event {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 30)
+    private String name;
+
+    @Column(nullable = true)
+    private LocalDateTime startDate;
+
+    @Column(nullable = true)
+    private LocalDateTime endDate;
+}
