+++ src/main/java/com/flab/s_market/domains/term/domain/SubTerm.java
@@ -1,0 +1,30 @@
+package com.flab.s_market.domains.term.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EmbeddedId;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MapsId;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class SubTerm {
+    @EmbeddedId
+    private SubTermId id;
+
+    @Column(nullable = false, length = 255)
+    private String content;
+
+    @MapsId("termId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "term_id")
+    private Term term;
+}
