+++ src/main/java/com/flab/s_market/domains/term/domain/Term.java
@@ -1,0 +1,33 @@
+package com.flab.s_market.domains.term.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Term {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 255)
+    private String url;
+
+    @Column(nullable = false, length = 80)
+    private String title;
+
+    @Column(nullable = false)
+    private Boolean isRequired;
+
+    @Column(nullable = false)
+    private Boolean additional;
+
+}
