+++ src/main/java/com/flab/s_market/domains/term/domain/UserSubTerm.java
@@ -1,0 +1,34 @@
+package com.flab.s_market.domains.term.domain;
+
+import com.flab.s_market.domains.user.domain.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.EmbeddedId;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MapsId;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserSubTerm {
+    @EmbeddedId
+    private UserSubTermId id;
+
+    @Column(nullable = false)
+    private LocalDateTime agreeDate;
+
+    @MapsId("userId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+
+}
