+++ src/main/java/com/flab/s_market/domains/term/domain/UserSubTermId.java
@@ -1,0 +1,30 @@
+package com.flab.s_market.domains.term.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinColumns;
+import jakarta.persistence.ManyToOne;
+import java.io.Serializable;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode
+public class UserSubTermId implements Serializable {
+    @Column(nullable = false)
+    private Long userId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumns({
+        @JoinColumn(name = "version", referencedColumnName = "version"),
+        @JoinColumn(name = "term_id", referencedColumnName = "term_id")
+    })
+    private SubTerm subTerm;
+}
