+++ src/main/java/com/flab/s_market/domains/term/repository/TermRepository.java
@@ -1,0 +1,10 @@
+package com.flab.s_market.domains.term.repository;
+
+import com.flab.s_market.domains.term.domain.Term;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TermRepository extends JpaRepository<Term, Long> {
+
+}
