+++ src/main/java/com/flab/s_market/domains/user/dto/response/AllTermDTO.java
@@ -1,0 +1,11 @@
+package com.flab.s_market.domains.user.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder @Getter
+public class AllTermDTO {
+    private TermsDTO main;
+    private TermsDTO additional;
+
+}
