+++ src/main/java/com/flab/s_market/domains/user/dto/response/DetailTermDTO.java
@@ -1,0 +1,11 @@
+package com.flab.s_market.domains.user.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder @Getter
+public class DetailTermDTO {
+    private String url;
+    private double version;
+    private String title;
+}
