+++ src/main/java/com/flab/s_market/domains/user/dto/response/TermDTO.java
@@ -1,0 +1,10 @@
+package com.flab.s_market.domains.user.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder @Getter
+public class TermDTO {
+    private String title;
+    private boolean isRequired;
+}
