+++ src/main/java/com/flab/s_market/domains/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.flab.s_market.domains.user.repository;
+
+import com.flab.s_market.domains.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+}
