--- src/main/java/com/flab/s_market/order/domain/Order.java
@@ -1,37 +1,0 @@
-package com.flab.s_market.order.domain;
-
-import com.flab.s_market.delivery.domain.Delivery;
-import com.flab.s_market.shopping.domain.Shopping;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "Orders")
-public class Order {
-    @Id @GeneratedValue
-    @Column(name = "order_id")
-    private Long id;
-    private LocalDateTime orderDate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "shopping_id")
-    private Shopping shopping;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "delivery_id")
-    private Delivery delivery;
-}
