--- src/main/java/com/flab/s_market/shopping/domain/Product.java
@@ -1,25 +1,0 @@
-package com.flab.s_market.shopping.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Product {
-    @Id @GeneratedValue
-    @Column(name = "product_id")
-    private Long id;
-
-    private int amount;
-    private int price;
-    private String name;
-
-}
