--- src/main/java/com/flab/s_market/shopping/domain/Shopping.java
@@ -1,28 +1,0 @@
-package com.flab.s_market.shopping.domain;
-
-import com.flab.s_market.user.domain.User;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Shopping {
-    @Id @GeneratedValue
-    @Column(name = "shopping_id")
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-}
