--- src/main/java/com/flab/s_market/user/dto/response/TermsDTO.java
@@ -1,13 +1,0 @@
-package com.flab.s_market.user.dto.response;
-
-import java.util.List;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder @Getter
-public class TermsDTO {
-    private String title;
-    private boolean isRequired;
-    private List<TermDTO> terms;
-
-}
