--- src/main/java/com/flab/s_market/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.flab.s_market.user.repository;
-
-import com.flab.s_market.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByEmail(String email);
-}
